lass es worken Release v5

+++ mainwindow.cpp
@@ -176,15 +176,17 @@ bool ModernItemDelegate::editorEvent(QEvent *event, QAbstractItemModel *model, c
 }
 
 // --- ModernButton ---
-ModernButton::ModernButton(QWidget *parent) : QToolButton(parent), m_scale(1.0f), m_accentColor(Qt::white) {
+// ÄNDERUNG: qreal statt float
+ModernButton::ModernButton(QWidget *parent) : QToolButton(parent), m_scale(1.0), m_accentColor(Qt::white) {
     m_anim = new QPropertyAnimation(this, "scale", this); m_anim->setDuration(150); m_anim->setEasingCurve(QEasingCurve::OutQuad);
     setCursor(Qt::PointingHandCursor); setStyleSheet("border: none; background: transparent;");
     setIconSize(QSize(24, 24));
 }
-void ModernButton::setScale(float s) { m_scale = s; update(); }
+// ÄNDERUNG: qreal
+void ModernButton::setScale(qreal s) { m_scale = s; update(); }
 void ModernButton::setAccentColor(QColor c) { m_accentColor = c; update(); }
-void ModernButton::enterEvent(QEnterEvent *event) { m_anim->stop(); m_anim->setEndValue(1.2f); m_anim->start(); QToolButton::enterEvent(event); }
-void ModernButton::leaveEvent(QEvent *event) { m_anim->stop(); m_anim->setEndValue(1.0f); m_anim->start(); QToolButton::leaveEvent(event); }
+void ModernButton::enterEvent(QEnterEvent *event) { m_anim->stop(); m_anim->setEndValue(1.2); m_anim->start(); QToolButton::enterEvent(event); }
+void ModernButton::leaveEvent(QEvent *event) { m_anim->stop(); m_anim->setEndValue(1.0); m_anim->start(); QToolButton::leaveEvent(event); }
 void ModernButton::paintEvent(QPaintEvent *) {
     QPainter p(this); p.setRenderHint(QPainter::Antialiasing); p.setRenderHint(QPainter::SmoothPixmapTransform);
     QIcon ic = icon(); if (ic.isNull()) return; int w = width(); int h = height();
